+++ go/tasks/plugins/k8s/sagemaker/builtin_training_test.go
@@ -1,0 +1,196 @@
+package sagemaker
+
+import (
+	"context"
+	"testing"
+
+	trainingjobController "github.com/aws/amazon-sagemaker-operator-for-k8s/controllers/trainingjob"
+	flyteIdlCore "github.com/lyft/flyteidl/gen/pb-go/flyteidl/core"
+	taskError "github.com/lyft/flyteplugins/go/tasks/errors"
+	pluginsCore "github.com/lyft/flyteplugins/go/tasks/pluginmachinery/core"
+
+	commonv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/common"
+
+	"github.com/lyft/flyteplugins/go/tasks/plugins/k8s/sagemaker/config"
+
+	stdConfig "github.com/lyft/flytestdlib/config"
+	"github.com/lyft/flytestdlib/config/viper"
+
+	trainingjobv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/trainingjob"
+	"github.com/aws/aws-sdk-go/service/sagemaker"
+	sagemakerIdl "github.com/lyft/flyteidl/gen/pb-go/flyteidl/plugins/sagemaker"
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_awsSagemakerPlugin_BuildResourceForTrainingJob(t *testing.T) {
+	// Default config does not contain a roleAnnotationKey -> expecting to get the role from default config
+	ctx := context.TODO()
+	defaultCfg := config.GetSagemakerConfig()
+	defer func() {
+		_ = config.SetSagemakerConfig(defaultCfg)
+	}()
+
+	t.Run("If roleAnnotationKey has a match, the role from the metadata should be fetched", func(t *testing.T) {
+		// Injecting a config which contains a matching roleAnnotationKey -> expecting to get the role from metadata
+		configAccessor := viper.NewAccessor(stdConfig.Options{
+			StrictMode:  true,
+			SearchPaths: []string{"testdata/config.yaml"},
+		})
+
+		err := configAccessor.UpdateConfig(context.TODO())
+		assert.NoError(t, err)
+
+		awsSageMakerTrainingJobHandler := awsSagemakerPlugin{TaskType: trainingJobTaskType}
+
+		tjObj := generateMockTrainingJobCustomObj(
+			sagemakerIdl.InputMode_FILE, sagemakerIdl.AlgorithmName_XGBOOST, "0.90", []*sagemakerIdl.MetricDefinition{},
+			sagemakerIdl.InputContentType_TEXT_CSV, 1, "ml.m4.xlarge", 25)
+		taskTemplate := generateMockTrainingJobTaskTemplate("the job x", tjObj)
+
+		trainingJobResource, err := awsSageMakerTrainingJobHandler.BuildResource(ctx, generateMockTrainingJobTaskContext(taskTemplate, false))
+		assert.NoError(t, err)
+		assert.NotNil(t, trainingJobResource)
+
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+		assert.Equal(t, "metadata_role", *trainingJob.Spec.RoleArn)
+	})
+
+	t.Run("If roleAnnotationKey does not have a match, the role from the config should be fetched", func(t *testing.T) {
+		// Injecting a config which contains a mismatched roleAnnotationKey -> expecting to get the role from the config
+		configAccessor := viper.NewAccessor(stdConfig.Options{
+			StrictMode: true,
+			// Use a different
+			SearchPaths: []string{"testdata/config2.yaml"},
+		})
+
+		err := configAccessor.UpdateConfig(context.TODO())
+		assert.NoError(t, err)
+
+		awsSageMakerTrainingJobHandler := awsSagemakerPlugin{TaskType: trainingJobTaskType}
+
+		tjObj := generateMockTrainingJobCustomObj(
+			sagemakerIdl.InputMode_FILE, sagemakerIdl.AlgorithmName_XGBOOST, "0.90", []*sagemakerIdl.MetricDefinition{},
+			sagemakerIdl.InputContentType_TEXT_CSV, 1, "ml.m4.xlarge", 25)
+		taskTemplate := generateMockTrainingJobTaskTemplate("the job y", tjObj)
+
+		trainingJobResource, err := awsSageMakerTrainingJobHandler.BuildResource(ctx, generateMockTrainingJobTaskContext(taskTemplate, false))
+		assert.NoError(t, err)
+		assert.NotNil(t, trainingJobResource)
+
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+		assert.Equal(t, "config_role", *trainingJob.Spec.RoleArn)
+	})
+
+	t.Run("In a custom training job we should see the FLYTE_SAGEMAKER_CMD being injected", func(t *testing.T) {
+		// Injecting a config which contains a mismatched roleAnnotationKey -> expecting to get the role from the config
+		configAccessor := viper.NewAccessor(stdConfig.Options{
+			StrictMode: true,
+			// Use a different
+			SearchPaths: []string{"testdata/config2.yaml"},
+		})
+
+		err := configAccessor.UpdateConfig(context.TODO())
+		assert.NoError(t, err)
+
+		awsSageMakerTrainingJobHandler := awsSagemakerPlugin{TaskType: trainingJobTaskType}
+
+		tjObj := generateMockTrainingJobCustomObj(
+			sagemakerIdl.InputMode_FILE, sagemakerIdl.AlgorithmName_XGBOOST, "0.90", []*sagemakerIdl.MetricDefinition{},
+			sagemakerIdl.InputContentType_TEXT_CSV, 1, "ml.m4.xlarge", 25)
+		taskTemplate := generateMockTrainingJobTaskTemplate("the job", tjObj)
+
+		trainingJobResource, err := awsSageMakerTrainingJobHandler.BuildResource(ctx, generateMockTrainingJobTaskContext(taskTemplate, false))
+		assert.NoError(t, err)
+		assert.NotNil(t, trainingJobResource)
+
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+		assert.Equal(t, "config_role", *trainingJob.Spec.RoleArn)
+
+		expectedHPs := []*commonv1.KeyValuePair{
+			{Name: "a", Value: "1"},
+			{Name: "b", Value: "2"},
+		}
+
+		assert.ElementsMatch(t,
+			func(kvs []*commonv1.KeyValuePair) []commonv1.KeyValuePair {
+				ret := make([]commonv1.KeyValuePair, 0, len(kvs))
+				for _, kv := range kvs {
+					ret = append(ret, *kv)
+				}
+				return ret
+			}(expectedHPs),
+			func(kvs []*commonv1.KeyValuePair) []commonv1.KeyValuePair {
+				ret := make([]commonv1.KeyValuePair, 0, len(kvs))
+				for _, kv := range kvs {
+					ret = append(ret, *kv)
+				}
+				return ret
+			}(trainingJob.Spec.HyperParameters))
+	})
+}
+
+func Test_awsSagemakerPlugin_GetTaskPhaseForTrainingJob(t *testing.T) {
+	ctx := context.TODO()
+	// Injecting a config which contains a mismatched roleAnnotationKey -> expecting to get the role from the config
+	configAccessor := viper.NewAccessor(stdConfig.Options{
+		StrictMode: true,
+		// Use a different
+		SearchPaths: []string{"testdata/config2.yaml"},
+	})
+
+	err := configAccessor.UpdateConfig(context.TODO())
+	assert.NoError(t, err)
+
+	awsSageMakerTrainingJobHandler := awsSagemakerPlugin{TaskType: trainingJobTaskType}
+
+	tjObj := generateMockTrainingJobCustomObj(
+		sagemakerIdl.InputMode_FILE, sagemakerIdl.AlgorithmName_XGBOOST, "0.90", []*sagemakerIdl.MetricDefinition{},
+		sagemakerIdl.InputContentType_TEXT_CSV, 1, "ml.m4.xlarge", 25)
+	taskTemplate := generateMockTrainingJobTaskTemplate("the job", tjObj)
+	taskCtx := generateMockTrainingJobTaskContext(taskTemplate, false)
+	trainingJobResource, err := awsSageMakerTrainingJobHandler.BuildResource(ctx, taskCtx)
+	assert.NoError(t, err)
+	assert.NotNil(t, trainingJobResource)
+
+	t.Run("ReconcilingTrainingJobStatus should lead to a retryable failure", func(t *testing.T) {
+
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+		trainingJob.Status.TrainingJobStatus = trainingjobController.ReconcilingTrainingJobStatus
+		phaseInfo, err := awsSageMakerTrainingJobHandler.getTaskPhaseForTrainingJob(ctx, taskCtx, trainingJob)
+		assert.Nil(t, err)
+		assert.Equal(t, phaseInfo.Phase(), pluginsCore.PhaseRetryableFailure)
+		assert.Equal(t, phaseInfo.Err().GetKind(), flyteIdlCore.ExecutionError_USER)
+		assert.Equal(t, phaseInfo.Err().GetCode(), trainingjobController.ReconcilingTrainingJobStatus)
+		assert.Equal(t, phaseInfo.Err().GetMessage(), "")
+	})
+	t.Run("TrainingJobStatusFailed should be a permanent failure", func(t *testing.T) {
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+		trainingJob.Status.TrainingJobStatus = sagemaker.TrainingJobStatusFailed
+
+		phaseInfo, err := awsSageMakerTrainingJobHandler.getTaskPhaseForTrainingJob(ctx, taskCtx, trainingJob)
+		assert.Nil(t, err)
+		assert.Equal(t, phaseInfo.Phase(), pluginsCore.PhasePermanentFailure)
+		assert.Equal(t, phaseInfo.Err().GetKind(), flyteIdlCore.ExecutionError_USER)
+		assert.Equal(t, phaseInfo.Err().GetCode(), sagemaker.TrainingJobStatusFailed)
+		assert.Equal(t, phaseInfo.Err().GetMessage(), "")
+	})
+	t.Run("TrainingJobStatusFailed should be a permanent failure", func(t *testing.T) {
+
+		trainingJob, ok := trainingJobResource.(*trainingjobv1.TrainingJob)
+		assert.True(t, ok)
+		trainingJob.Status.TrainingJobStatus = sagemaker.TrainingJobStatusStopped
+
+		phaseInfo, err := awsSageMakerTrainingJobHandler.getTaskPhaseForTrainingJob(ctx, taskCtx, trainingJob)
+		assert.Nil(t, err)
+		assert.Equal(t, phaseInfo.Phase(), pluginsCore.PhaseRetryableFailure)
+		assert.Equal(t, phaseInfo.Err().GetKind(), flyteIdlCore.ExecutionError_USER)
+		assert.Equal(t, phaseInfo.Err().GetCode(), taskError.DownstreamSystemError)
+		// We have a default message for TrainingJobStatusStopped
+		assert.Equal(t, phaseInfo.Err().GetMessage(), "Training Job Stopped")
+	})
+}
